fix:updated failing tests
