+++ src/test/resources/org/dvsa/testing/framework/features/SelfServe/Accessibility/manage-vehicle-licence-accessibility.feature
@@ -1,4 +1,3 @@
-@ss_regression
 Feature: Check that manage vehicle journey is accessible
 
   Background:
